+++ src/main/java/com/example/hipreader/auth/repository/RefreshTokenRepository.java
@@ -1,10 +1,10 @@
-package com.example.hipreader.domain.refreshtoken.repository;
+package com.example.hipreader.auth.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.repository.CrudRepository;
 
-import com.example.hipreader.domain.refreshtoken.entity.RefreshToken;
+import com.example.hipreader.auth.entity.RefreshToken;
 
 public interface RefreshTokenRepository extends CrudRepository<RefreshToken, Long> {
 	Optional<RefreshToken> findByRefreshToken(String token);
