+++ src/main/java/com/example/hipreader/auth/service/AuthService.java
@@ -5,7 +5,6 @@ import static com.example.hipreader.common.exception.ErrorCode.*;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.server.ResponseStatusException;
 
 import com.example.hipreader.auth.dto.request.SigninRequestDto;
 import com.example.hipreader.auth.dto.request.SignupRequestDto;
@@ -14,8 +13,8 @@ import com.example.hipreader.auth.dto.response.SignupResponseDto;
 import com.example.hipreader.common.exception.BadRequestException;
 import com.example.hipreader.common.exception.NotFoundException;
 import com.example.hipreader.common.util.JwtUtil;
-import com.example.hipreader.domain.refreshtoken.entity.RefreshToken;
-import com.example.hipreader.domain.refreshtoken.repository.RefreshTokenRepository;
+import com.example.hipreader.auth.entity.RefreshToken;
+import com.example.hipreader.auth.repository.RefreshTokenRepository;
 import com.example.hipreader.domain.user.entity.User;
 import com.example.hipreader.domain.user.gender.Gender;
 import com.example.hipreader.domain.user.repository.UserRepository;
